:recycle: refactor: move nextauth schema to `src/database/server/schemas`
